+++ src/main/java/com/atlauncher/data/minecraft/loaders/forge/ATLauncherApiForgeVersion.java
@@ -17,8 +17,18 @@
  */
 package com.atlauncher.data.minecraft.loaders.forge;
 
+import com.google.gson.annotations.SerializedName;
+
 public class ATLauncherApiForgeVersion {
     public String version;
     public boolean recommended;
-    public String raw_version;
+
+    @SerializedName("raw_version")
+    public String rawVersion;
+
+    @SerializedName("installer_sha1_hash")
+    public String installerSha1Hash;
+
+    @SerializedName("installer_size")
+    public Long installerSize;
 }